Move showemotes module
