+++ pajbot/modules/__init__.py
@@ -28,6 +28,7 @@ from pajbot.modules.chatters_refresh import ChattersRefreshModule
 from pajbot.modules.clr_overlay import CLROverlayModule
 from pajbot.modules.clr_overlay.emotecombo import EmoteComboModule
 from pajbot.modules.clr_overlay.emotesonscreen import EmotesOnScreenModule
+from pajbot.modules.clr_overlay.showemote import ShowEmoteModule
 from pajbot.modules.deck import DeckModule
 from pajbot.modules.duel import DuelModule
 from pajbot.modules.dummy import DummyModule
@@ -69,7 +70,6 @@ from pajbot.modules.quests.winraffle import WinRaffleQuestModule
 from pajbot.modules.raffle import RaffleModule
 from pajbot.modules.repspam import RepspamModule
 from pajbot.modules.roulette import RouletteModule
-from pajbot.modules.showemote import ShowEmoteModule
 from pajbot.modules.slotmachine import SlotMachineModule
 from pajbot.modules.subscriber_fetch import SubscriberFetchModule
 from pajbot.modules.top import TopModule
